dns_manager/requirements.txt
psycopg2-binary
requests
python-dotenv
schedule
dnspython
